feat: add validate field for templateValue secrets
Replicator's webhook will deny an admission if a referenced secret does
not exist at the time of apply.

Sometimes however, you want to reference a secret that should eventually
be provisioned by another controller. While you could easily solve this by
just ensuring that the secret exists before applying the RepConf, this
isn't always an option, e.g. when both of these resources are applied
through the same Helm Chart.

This is opt-in with a new field to avoid breaking compatibility
and assumptions for existing RepConfs where you generally want to know if
a referenced secret exists or not.

Replicator's reconcile-loop should retry until the secret
exists and the replication config is successfully synchronized.